requirements.txt
action-msgs==2.0.3
action-tutorials-interfaces==0.33.7
action-tutorials-py==0.33.7
actionlib-msgs==5.3.6
actuator-msgs==0.0.1
ament-cmake-test==2.5.4
ament-copyright==0.17.3
ament-cppcheck==0.17.3
ament-cpplint==0.17.3
ament-flake8==0.17.3
ament-index-python==1.8.1
ament-lint==0.17.3
ament-lint-cmake==0.17.3
ament-package==0.16.4
ament-pep257==0.17.3
ament-uncrustify==0.17.3
ament-xmllint==0.17.3
angles==1.16.1
annotated-types==0.7.0
anyio==4.12.0
builtin-interfaces==2.0.3
certifi==2025.11.12
composition-interfaces==2.0.3
control-msgs==5.5.0
controller-manager==4.39.2
controller-manager-msgs==4.39.2
cv-bridge==4.1.0
demo-nodes-py==0.33.7
diagnostic-msgs==5.3.6
diagnostic-updater==4.2.6
distro==1.9.0
domain-coordinator==0.12.0
example-interfaces==0.12.0
examples-rclpy-executors==0.19.7
examples-rclpy-minimal-action-client==0.19.7
examples-rclpy-minimal-action-server==0.19.7
examples-rclpy-minimal-client==0.19.7
examples-rclpy-minimal-publisher==0.19.7
examples-rclpy-minimal-service==0.19.7
examples-rclpy-minimal-subscriber==0.19.7
faiss-cpu==1.13.0
generate-parameter-library-py==0.5.0
geometry-msgs==5.3.6
gps-msgs==2.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
image-geometry==4.1.0
interactive-markers==2.5.5
jiter==0.12.0
joint-state-publisher==2.4.0
laser-geometry==2.7.2
launch==3.4.9
launch-ros==0.26.10
launch-testing==3.4.9
launch-testing-ros==0.26.10
launch-xml==3.4.9
launch-yaml==3.4.9
lifecycle-msgs==2.0.3
logging-demo==0.33.7
map-msgs==2.4.1
# Editable install with no version control (mb==0.0.0)
-e /home/lvanegas/mb_ws/build/mb
# Editable install with no version control (mb-ign==0.0.0)
-e /home/lvanegas/mb_ws/build/mb_ign
message-filters==4.11.9
nav-msgs==5.3.6
numpy==2.3.5
openai==2.8.1
osrf-pycommon==2.1.7
packaging==25.0
pal-statistics==2.7.0
pal-statistics-msgs==2.7.0
pcl-msgs==1.0.0
pendulum-msgs==0.33.7
pydantic==2.12.5
pydantic_core==2.41.5
PyPDF2==3.0.1
python-qt-binding==2.2.2
qt-dotgraph==2.7.5
qt-gui==2.7.5
qt-gui-cpp==2.7.5
qt-gui-py-common==2.7.5
quality-of-service-demo-py==0.33.7
rcl-interfaces==2.0.3
rclpy==7.1.6
rcutils==6.7.4
resource-retriever==3.4.4
rmw-dds-common==3.1.0
ros-gz-bridge==1.0.16
ros-gz-interfaces==1.0.16
ros-gz-sim==1.0.16
ros2action==0.32.6
ros2bag==0.26.9
ros2bag-mcap-cli==0.26.9
ros2bag-sqlite3-cli==0.26.9
ros2cli==0.32.6
ros2component==0.32.6
ros2doctor==0.32.6
ros2interface==0.32.6
ros2launch==0.26.10
ros2lifecycle==0.32.6
ros2multicast==0.32.6
ros2node==0.32.6
ros2param==0.32.6
ros2pkg==0.32.6
ros2run==0.32.6
ros2service==0.32.6
ros2topic==0.32.6
rosapi==2.4.1
rosapi-msgs==2.4.1
rosbag2-interfaces==0.26.9
rosbag2-py==0.26.9
rosbridge-library==2.3.0
rosbridge-msgs==2.4.1
rosbridge-server==2.4.1
rosgraph-msgs==2.0.3
rosidl-adapter==4.6.6
rosidl-cli==4.6.6
rosidl-cmake==4.6.6
rosidl-generator-c==4.6.6
rosidl-generator-cpp==4.6.6
rosidl-generator-py==0.22.2
rosidl-generator-type-description==4.6.6
rosidl-parser==4.6.6
rosidl-pycommon==4.6.6
rosidl-runtime-py==0.13.1
rosidl-typesupport-c==3.2.2
rosidl-typesupport-cpp==3.2.2
rosidl-typesupport-fastrtps-c==3.6.2
rosidl-typesupport-fastrtps-cpp==3.6.2
rosidl-typesupport-introspection-c==4.6.6
rosidl-typesupport-introspection-cpp==4.6.6
rpyutils==0.4.2
rqt-action==2.2.0
rqt-bag==1.5.5
rqt-bag-plugins==1.5.5
rqt-console==2.2.1
rqt-graph==1.5.5
rqt-gui==1.6.2
rqt-gui-py==1.6.2
rqt-msg==1.5.1
rqt-plot==1.4.4
rqt-publisher==1.7.2
rqt-py-common==1.6.2
rqt-py-console==1.2.3
rqt-reconfigure==1.6.2
rqt-service-caller==1.2.1
rqt-shell==1.2.2
rqt-srv==1.2.2
rqt-topic==1.7.4
sensor-msgs==5.3.6
sensor-msgs-py==5.3.6
service-msgs==2.0.3
setuptools==80.9.0
shape-msgs==5.3.6
sniffio==1.3.1
sros2==0.13.4
statistics-msgs==2.0.3
std-msgs==5.3.6
std-srvs==5.3.6
stereo-msgs==5.3.6
teleop-twist-keyboard==2.4.1
tf2-geometry-msgs==0.36.16
tf2-kdl==0.36.16
tf2-msgs==0.36.16
tf2-py==0.36.16
tf2-ros-py==0.36.16
tf2-sensor-msgs==0.36.16
tf2-tools==0.36.16
theora-image-transport==4.0.6
topic-monitor==0.33.7
tqdm==4.67.1
trajectory-msgs==5.3.6
turtlesim==1.8.3
type-description-interfaces==2.0.3
typing-inspection==0.4.2
typing_extensions==4.15.0
unique-identifier-msgs==2.5.0
vision-msgs==4.1.1
visualization-msgs==5.3.6
wheel==0.45.1
xacro==2.1.1
